# [static runtime][wip] Integrate NNC with hash-based caching mechanism