--- a/torch/csrc/jit/runtime/static/impl.cpp
+++ b/torch/csrc/jit/runtime/static/impl.cpp
@@ -104,110 +104,36 @@ StaticRuntime::StaticRuntime(const torch::jit::Module& m)
     }
   }
 
-  // fill constant_table_ and operator_table_
+  // fill workspace_ with constants
   for (Node* node : graph_->nodes()) {
-    switch (node->kind()) {
-      case prim::Constant:
-        CHECK(node->output()->type()->kind() != FunctionType::Kind);
-        constant_table_[node->output()] = toIValue(node->output()).value();
-        break;
-      case prim::ListConstruct:
-        nodes_.emplace_back(node, nullptr);
-        break;
-      case prim::TupleConstruct:
-        nodes_.emplace_back(node, nullptr);
-        break;
-      default: {
-        const Operator& op = node->getOperator();
-        CHECK(op.hasOperation());
-        nodes_.emplace_back(node, op.getOperation(node));
-      }
-    }
-  }
-}
-
-void StaticRuntime::getInputIValues(
-    Node* node,
-    const ConstantMap& ws,
-    std::vector<IValue>& stack) const {
-  const size_t size = node->inputs().size();
-  stack.reserve(size);
-  for (size_t i = 0; i < size; i++) {
-    Value* v = node->inputs()[i];
-    auto f = constant_table_.find(v);
-    if (f == constant_table_.end()) {
-      auto f_ws = ws.find(v);
-      TORCH_CHECK(
-          f_ws != ws.end(),
-          "Workspace does not contain Value ",
-          v->debugName());
-      stack.emplace_back(f_ws->second);
+    if (node->kind() == prim::Constant) {
+      CHECK(node->output()->type()->kind() != FunctionType::Kind);
+      workspace_[node->output()] = toIValue(node->output()).value();
     } else {
-      stack.emplace_back(f->second);
+      nodes_.emplace_back(node);
     }
-  }
-}
-
-void StaticRuntime::runNodes(ConstantMap& workspace) const {
-  std::vector<IValue> stack;
-  for (const auto& p : nodes_) {
-    Node* node = p.first;
-    const Operation& op = p.second;
-    getInputIValues(node, workspace, stack);
-    VLOG(1) << node->kind().toDisplayString();
-
-    switch (node->kind()) {
-      case prim::ListConstruct: {
-        listConstruct(
-            stack,
-            node->output()->type()->expect<ListType>(),
-            node->inputs().size());
-      } break;
-      case prim::TupleConstruct: {
-        bool named =
-            node->output()->type()->expect<TupleType>()->name().has_value();
-        if (named) {
-          namedTupleConstruct(
-              stack,
-              node->output()->type()->expect<TupleType>(),
-              node->inputs().size());
-        } else {
-          tupleConstruct(stack, node->inputs().size());
-        }
-      } break;
-      default: {
-        DCHECK(op);
-        op(&stack);
-        break;
-      }
-    }
-
-    DCHECK_EQ(stack.size(), node->outputs().size());
-    for (auto i = 0; i < node->outputs().size(); i++) {
-      workspace[node->outputs()[i]] = stack[i];
-    }
-    stack.clear();
   }
 }
 
 std::vector<at::Tensor> StaticRuntime::run(
-    const std::vector<at::Tensor>& inps) const {
+    const std::vector<at::Tensor>& inps) {
   // Container for inputs, outputs, and activations (excluding parameters)
-  ConstantMap workspace_;
 
   int start = 0;
   if (graph_->inputs().size() != inps.size()) {
     start = 1;
     CHECK_EQ(graph_->inputs().size(), inps.size() + 1);
     CHECK((graph_->inputs().at(0)->type()->is_module()));
-    workspace_.emplace(graph_->inputs()[0], module_._ivalue());
+    workspace_[graph_->inputs()[0]] = module_._ivalue();
   }
 
   for (size_t i = 0; i < inps.size(); i++) {
-    workspace_.emplace(graph_->inputs()[i + start], inps[i]);
+    workspace_[graph_->inputs()[i + start]] = inps[i];
   }
 
-  runNodes(workspace_);
+  for (const auto& n : nodes_) {
+    n.run(workspace_);
+  }
 
   std::vector<at::Tensor> out;
   for (Value* output : graph_->outputs()) {
@@ -223,5 +149,61 @@ std::vector<at::Tensor> StaticRuntime::run(
   }
   return out;
 }
+
+ProcessedNode::ProcessedNode(Node* node) : node_(node) {
+  if (node->kind() != prim::ListConstruct &&
+      node->kind() != prim::TupleConstruct) {
+    const Operator& op = node->getOperator();
+    CHECK(op.hasOperation());
+    op_ = op.getOperation(node);
+  }
+}
+
+void ProcessedNode::run(StaticRuntime::ConstantMap& workspace) const {
+  if (use_stack_) {
+    std::vector<IValue> stack;
+    const size_t size = node_->inputs().size();
+    stack.reserve(size);
+    for (size_t i = 0; i < size; i++) {
+      Value* v = node_->inputs()[i];
+      auto f = workspace.find(v);
+      TORCH_CHECK(
+          f != workspace.end(),
+          "Workspace does not contain Value ",
+          v->debugName());
+      stack.emplace_back(f->second);
+    }
+    if (op_) {
+      (*op_)(&stack);
+    } else {
+      if (node_->kind() == prim::ListConstruct) {
+        listConstruct(
+            stack,
+            node_->output()->type()->expect<ListType>(),
+            node_->inputs().size());
+      } else if (node_->kind() == prim::TupleConstruct) {
+        bool named =
+            node_->output()->type()->expect<TupleType>()->name().has_value();
+        if (named) {
+          namedTupleConstruct(
+              stack,
+              node_->output()->type()->expect<TupleType>(),
+              node_->inputs().size());
+        } else {
+          tupleConstruct(stack, node_->inputs().size());
+        }
+      } else {
+        TORCH_CHECK(0, "Unhandled operation!", node_->kind().toQualString());
+      }
+    }
+    DCHECK_EQ(stack.size(), node_->outputs().size());
+    for (auto i = 0; i < node_->outputs().size(); i++) {
+      workspace[node_->outputs()[i]] = stack[i];
+    }
+  } else {
+    TORCH_CHECK(0, "Non-stack execution not yet implemented");
+  }
+}
+
 } // namespace jit
 } // namespace torch