--- a/torch/csrc/jit/runtime/static/impl.cpp
+++ b/torch/csrc/jit/runtime/static/impl.cpp
@@ -1,9 +1,10 @@
-#include <torch/csrc/jit/runtime/static/impl.h>
+#include <ATen/core/interned_strings.h>
 #include <ATen/core/op_registration/op_registration.h>
 #include <torch/csrc/jit/passes/canonicalize.h>
 #include <torch/csrc/jit/passes/freeze_module.h>
 #include <torch/csrc/jit/passes/remove_mutation.h>
 #include <torch/csrc/jit/passes/subgraph_rewrite.h>
+#include <torch/csrc/jit/runtime/static/impl.h>
 #include <torch/csrc/jit/runtime/static/ops.h>
 #include <torch/csrc/jit/runtime/vararg_functions.h>
 
@@ -123,7 +124,7 @@ void ProcessedNode::run(StaticRuntime::ConstantMap& workspace) const {
       stack.emplace_back(f->second);
     }
     if (op_) {
-      (*op_)(&stack);
+      op_->operator()(&stack);
     } else {
       if (node_->kind() == prim::ListConstruct) {
         listConstruct(
@@ -150,7 +151,7 @@ void ProcessedNode::run(StaticRuntime::ConstantMap& workspace) const {
       workspace[node_->outputs()[i]] = stack[i];
     }
   } else {
-    (*fn_)(workspace);
+    fn_->operator()(workspace);
   }
 }