… benchmarks/static_runtime/CMakeLists.txt
… benchmarks/static_runtime/test_static_runtime.cc
… caffe2/CMakeLists.txt
… cmake/Summary.cmake
… test/test_static_runtime.py